% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri_workflowInit.R
\name{ri_workflowInit}
\alias{ri_workflowInit}
\title{Initialises the workflow}
\usage{
ri_workflowInit(script, workdir = getwd())
}
\arguments{
\item{script}{name of current script}

\item{workdir}{name of working directory which contains script}
}
\value{
nothing

This functions performs checks to assure that the packages git2r,
ricmd, and data file are loaded. It also checks if the script
exists and is part of a git repository. If any of the checks fails,
the function returns an error.

After the checks this function creates the project meta data. This
meta data is used when storing data objects into iRODS.
}
\description{
Each workflow must be intialised before it can be used. This
function performs some checks and sets the project meta data
}
